Upgrade node-sass to sass (Dart Sass) (#11)
node-sass was a huge pain in the ass dependency. It was deprecated,
which caused us to pull in a bunch of vulnerable packages, and it also
required Python 2 to build. Trying to install Python2 nowadays sucks.
So I bit the bullet and upgraded node-sass to use the new (Dart) sass
package, as recommended at https://www.npmjs.com/package/node-sass

Specifically, this was performed to address
https://github.com/copia-automation/react-diff-viewer/security/dependabot/45